requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-dotenv
pymupdf
bcrypt
